[MOD/#87] 파일 위치 수정
