+++ presentation/src/main/java/com/going/presentation/dashboard/triplist/ongoing/OngoingViewHolder.kt
@@ -1,24 +1,29 @@
-package com.going.presentation.dashboard.triplist
+package com.going.presentation.dashboard.triplist.ongoing
 
 import androidx.recyclerview.widget.RecyclerView
 import com.going.domain.entity.response.DashBoardModel.DashBoardTripModel
 import com.going.presentation.R
-import com.going.presentation.databinding.ItemDashBoardCompletedBinding
+import com.going.presentation.databinding.ItemDashBoardOngoingBinding
 import com.going.ui.extension.setOnSingleClickListener
 
-class CompletedViewHolder(
-    val binding: ItemDashBoardCompletedBinding,
-    private val listener: CompletedAdapter.OnDashBoardSelectedListener
+class OngoingViewHolder(
+    val binding: ItemDashBoardOngoingBinding,
+    private val listener: OngoingAdapter.OnDashBoardSelectedListener
 ) : RecyclerView.ViewHolder(binding.root) {
 
     fun onBind(item: DashBoardTripModel) {
-
         binding.run {
             tvDashboardTripTitle.text = item.title
             tvDashboardDateStart.text = item.startDate
             tvDashboardDateEnd.text = item.endDate
-            tvDashboardDeadlineCompleted.text =
-                itemView.context.getString(R.string.dashboard_tv_completed_trip)
+
+            if (item.day <= 0) {
+                tvDashboardDeadline.text =
+                    itemView.context.getString(R.string.dashboard_tv_traveling)
+            } else {
+                tvDashboardDeadline.text =
+                    itemView.context.getString(R.string.dashboard_tv_deadline, item.day)
+            }
 
             layoutDashboard.setOnSingleClickListener {
                 listener.onDashBoardSelectedListener(item)
